[TD-344] 유저가 제보한 가게, 작성한 리뷰, 보유한 메달 갯수 레디스에 캐시
