+++ threedollar-application-service/src/main/java/com/depromeet/threedollar/application/service/faq/FaqService.kt
@@ -1,8 +1,8 @@
 package com.depromeet.threedollar.application.service.faq
 
-import com.depromeet.threedollar.application.config.cache.CacheType
 import com.depromeet.threedollar.application.service.faq.dto.request.RetrieveFaqsRequest
 import com.depromeet.threedollar.application.service.faq.dto.response.FaqResponse
+import com.depromeet.threedollar.domain.config.cache.CacheType.CacheKey.FAQS
 import com.depromeet.threedollar.domain.domain.faq.FaqRepository
 import org.springframework.cache.annotation.Cacheable
 import org.springframework.stereotype.Service
@@ -13,7 +13,7 @@ class FaqService(
     private val faqRepository: FaqRepository
 ) {
 
-    @Cacheable(key = "#request.category?:'ALL'", value = [CacheType.CacheKey.FAQS])
+    @Cacheable(key = "#request.category?:'ALL'", value = [FAQS])
     @Transactional(readOnly = true)
     fun retrieveFaqsByCategory(request: RetrieveFaqsRequest): List<FaqResponse> {
         return faqRepository.findAllByCategory(request.category).asSequence()
