+++ threedollar-application-service/src/main/java/com/depromeet/threedollar/application/service/popup/PopupService.kt
@@ -1,8 +1,8 @@
 package com.depromeet.threedollar.application.service.popup
 
-import com.depromeet.threedollar.application.config.cache.CacheType
 import com.depromeet.threedollar.application.service.popup.dto.request.GetActivatedPopupsRequest
 import com.depromeet.threedollar.application.service.popup.dto.response.PopupResponse
+import com.depromeet.threedollar.domain.config.cache.CacheType.CacheKey.POPUP
 import com.depromeet.threedollar.domain.domain.popup.PopupRepository
 import org.springframework.cache.annotation.Cacheable
 import org.springframework.stereotype.Service
@@ -14,7 +14,7 @@ class PopupService(
     private val popupRepository: PopupRepository
 ) {
 
-    @Cacheable(key = "#request.platform", value = [CacheType.CacheKey.POPUP])
+    @Cacheable(key = "#request.platform", value = [POPUP])
     @Transactional(readOnly = true)
     fun getActivatedPopups(request: GetActivatedPopupsRequest): List<PopupResponse> {
         return popupRepository.findActivatedPopupsByPlatform(request.platform, LocalDateTime.now())
